Keep silence ids on the object
Bosun does something stupid and generates redis ids based on the SHA1 hash of
the start and end times (both parsed from a human-readable, rounded string), the
alert name, and the tag set (an unordered hash map). This causes some silences
to return with different IDs than they really are, causing deletes to not work.

With this change, the silence object keeps the ID as a string.